fix compiler warnings about narrowing conversions and non-copyable types.
(cherry picked from commit 31a93414b84d85b8185902c0fb994c1d9ec88bd0)

+++ test/generic/datagram_protocol.cpp
@@ -73,7 +73,7 @@ void test()
     dp::socket socket2(ios, dp(af_inet, ipproto_udp));
     dp::socket socket3(ios, dp::endpoint());
 #if !defined(BOOST_ASIO_WINDOWS_RUNTIME)
-    int native_socket1 = ::socket(af_inet, sock_dgram, 0);
+    intptr_t native_socket1 = ::socket(af_inet, sock_dgram, 0);
     dp::socket socket4(ios, dp(af_inet, ipproto_udp), native_socket1);
 #endif // !defined(BOOST_ASIO_WINDOWS_RUNTIME)
 
@@ -105,9 +105,9 @@ void test()
     socket1.open(dp(af_inet, ipproto_udp), ec);
 
 #if !defined(BOOST_ASIO_WINDOWS_RUNTIME)
-    int native_socket2 = ::socket(af_inet, sock_dgram, 0);
+    intptr_t native_socket2 = ::socket(af_inet, sock_dgram, 0);
     socket1.assign(dp(af_inet, ipproto_udp), native_socket2);
-    int native_socket3 = ::socket(af_inet, sock_dgram, 0);
+    intptr_t native_socket3 = ::socket(af_inet, sock_dgram, 0);
     socket1.assign(dp(af_inet, ipproto_udp), native_socket3, ec);
 #endif // !defined(BOOST_ASIO_WINDOWS_RUNTIME)
