+++ test/ip/tcp.cpp
@@ -198,7 +198,7 @@ void test()
     ip::tcp::socket socket4(ios, ip::tcp::endpoint(ip::tcp::v4(), 0));
     ip::tcp::socket socket5(ios, ip::tcp::endpoint(ip::tcp::v6(), 0));
 #if !defined(BOOST_ASIO_WINDOWS_RUNTIME)
-    int native_socket1 = ::socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
+    intptr_t native_socket1 = ::socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
     ip::tcp::socket socket6(ios, ip::tcp::v4(), native_socket1);
 #endif // !defined(BOOST_ASIO_WINDOWS_RUNTIME)
 
@@ -234,9 +234,9 @@ void test()
     socket1.open(ip::tcp::v6(), ec);
 
 #if !defined(BOOST_ASIO_WINDOWS_RUNTIME)
-    int native_socket2 = ::socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
+    intptr_t native_socket2 = ::socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
     socket1.assign(ip::tcp::v4(), native_socket2);
-    int native_socket3 = ::socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
+    intptr_t native_socket3 = ::socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
     socket1.assign(ip::tcp::v4(), native_socket3, ec);
 #endif // !defined(BOOST_ASIO_WINDOWS_RUNTIME)
 
@@ -657,7 +657,7 @@ void test()
     ip::tcp::acceptor acceptor4(ios, ip::tcp::endpoint(ip::tcp::v4(), 0));
     ip::tcp::acceptor acceptor5(ios, ip::tcp::endpoint(ip::tcp::v6(), 0));
 #if !defined(BOOST_ASIO_WINDOWS_RUNTIME)
-    int native_acceptor1 = ::socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
+    intptr_t native_acceptor1 = ::socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
     ip::tcp::acceptor acceptor6(ios, ip::tcp::v4(), native_acceptor1);
 #endif // !defined(BOOST_ASIO_WINDOWS_RUNTIME)
 
@@ -685,9 +685,9 @@ void test()
     acceptor1.open(ip::tcp::v6(), ec);
 
 #if !defined(BOOST_ASIO_WINDOWS_RUNTIME)
-    int native_acceptor2 = ::socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
+    intptr_t native_acceptor2 = ::socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
     acceptor1.assign(ip::tcp::v4(), native_acceptor2);
-    int native_acceptor3 = ::socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
+    intptr_t native_acceptor3 = ::socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
     acceptor1.assign(ip::tcp::v4(), native_acceptor3, ec);
 #endif // !defined(BOOST_ASIO_WINDOWS_RUNTIME)
